+++ src/sketches/PlatonicSolids/PlatonicSolidsSketch.ts
@@ -6,7 +6,7 @@ import { OCTAHEDRON_GEOMETRY } from '@/meshes/Octahedron.js'
 import { DODECAHEDRON_GEOMETRY } from '@/meshes/Dodecahedron.js'
 import { ICOSAHEDRON_GEOMETRY } from '@/meshes/Icosahedron.js'
 import { encode_stl, encode_stl_file, mesh_to_triangle_soup } from '@/stl/encode_stl.js'
-import { CUBE_GEOMETRY } from '@/meshes/CubeGeometry.js'
+import { CUBE_GEOMETRY } from '@/meshes/Cube.js'
 import { download_file } from '@/core/download_file.js'
 import type { Mesh } from '@/meshes/Mesh.js'
 import type { InputSystem } from '@/input/InputSystem.js'
